+++ Backend/ATMMiniproject/ATMMiniproject/Models/Account.cs
@@ -1,4 +1,5 @@
 ﻿using ATM_MiniProject.Models;
+using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;
 
 namespace ATM_MiniProject.Context
@@ -6,8 +7,8 @@ namespace ATM_MiniProject.Context
     public class Account
     {
         [Key]
-        public int AcctId { get; set; }    
-        public float Balance { get; set; }
+        public int AcctId { get; set; }
+        public int Balance { get; set; }
          
         public IEnumerable<ATMTransaction> Transactions { get; set; } = new List<ATMTransaction>();
         public DebitCardDetails DebitCardDetails { get; set; }
